Just a proof of concept on how it'd be possible to fix up the grid system to allow for priority modifications from within filters (i.e. the user filter always taking precedence over others).
Obviously having the FilterResult in IFilterGUI's passesFilter method would be nicer, but I believe that's where other mods using the API would break, so I did it in IMultiFilter and added a special check for it in the SortingGrid. Obviously in 1.8 we could move it to the general API and have all passesFilter functions use the FilterResult class.

Related to #258