Power Limiter and Sorting Interface now make it's connected side more obvious - the side textures are now directed towards the facing side.
This was mostly done to solve an issue I was having with placing a Power Limiter on the ground, not knowing it had to be specifically connected to something, and I couldn't ever see the facing side since it was facing the ground so I thought they were broken altogether. It also looks a bit cooler I think.

Fixes #307